updated code to use redis
